Multiple reporters support
